Bump kaminpar submodule to up to date version

+++ include/kaminpar_wrapper.h
@@ -1,9 +1,10 @@
 #pragma once
 #include <memory>
 #include <vector>
+#include <optional>
 #include "rust/cxx.h"
 
-#include "kaminpar/vendor/KaMinPar/library/libkaminpar.h"
+#include "kaminpar/vendor/KaMinPar/kaminpar-shm/kaminpar.h"
 
 class PartitionerBuilder
 {
@@ -12,7 +13,7 @@ public:
     void set_threads(int threads);
     void set_epsilon(double epsilon);
     void set_seed(uint64_t seed);
-    void set_edge_weights(rust::Vec<int64_t> edge_weights);
+    void set_edge_weights(rust::Vec<int32_t> edge_weights);
     void set_node_weights(rust::Vec<int32_t> node_weights);
     std::unique_ptr<std::vector<uint32_t>> partition(rust::Vec<uint64_t> nodes, rust::Vec<uint32_t> edges, uint32_t num_partitions);
 
@@ -21,7 +22,7 @@ private:
     double m_epsilon;
     uint64_t m_seed;
     std::optional<rust::Vec<int32_t>> m_node_weights;
-    std::optional<rust::Vec<int64_t>> m_edge_weights;
+    std::optional<rust::Vec<int32_t>> m_edge_weights;
 };
 
 std::unique_ptr<PartitionerBuilder> new_partition_builder();
